feat: moved skill chips into expandable studentchapter views
