+++ components/ChapterHeader.tsx
@@ -5,7 +5,6 @@ import { Chip, CircularProgress, IconButton, Typography } from "@mui/material";
 import dayjs from "dayjs";
 import { ReactNode } from "react";
 import { graphql, useFragment } from "react-relay";
-import { LightTooltip } from "./LightTooltip";
 
 export function stringToColor(string: string): string {
   let hash = 0;
@@ -66,67 +65,6 @@ export function ChapterHeader({
     _chapter
   );
 
-  function getReadableTextColor(backgroundColor: String) {
-    const hex = backgroundColor.replace("#", "");
-
-    const r = parseInt(hex.substring(0, 2), 16);
-    const g = parseInt(hex.substring(2, 4), 16);
-    const b = parseInt(hex.substring(4, 6), 16);
-
-    const brightness = (r * 299 + g * 587 + b * 114) / 1000;
-    return brightness > 125 ? "#000000" : "#FFFFFF";
-  }
-
-  const skillCategoryMap = new Map<string, string[]>();
-
-  chapter.skills
-    .filter((c) => c !== null)
-    .forEach((c) => {
-      if (!skillCategoryMap.has(c!.skillCategory)) {
-        skillCategoryMap.set(c!.skillCategory, []);
-      }
-      skillCategoryMap.get(c!.skillCategory)!.push(c!.skillName);
-    });
-
-  for (const key of skillCategoryMap.keys()) {
-    skillCategoryMap.get(key)!.sort();
-  }
-
-  const skillChips = Array.from(skillCategoryMap.entries())
-    .sort()
-    .map(([category, skillNames], index) => (
-      <LightTooltip
-        key={category}
-        title={
-          <>
-            <p>
-              <strong>{category + ":"}</strong>
-            </p>
-            <ul className="list-disc pl-6">
-              {[...new Set(skillNames)].map((skillName, index) => (
-                <li key={index}>{skillName}</li>
-              ))}
-            </ul>
-          </>
-        }
-      >
-        <Chip
-          key={index}
-          sx={{
-            fontSize: "0.75rem",
-            height: "1.5rem",
-            maxWidth: "250px",
-            textOverflow: "ellipsis",
-            whiteSpace: "nowrap",
-            overflow: "hidden",
-            backgroundColor: stringToColor(category),
-            color: getReadableTextColor(stringToColor(category)),
-          }}
-          label={category}
-        />
-      </LightTooltip>
-    ));
-
   return (
     <div
       className="flex flex-row justify-start items-center p-6 rounded-3xl gap-16"
@@ -134,7 +72,7 @@ export function ChapterHeader({
     >
       <div className="flex flex-row items-center justify-center flex-grow">
         {(expandable === undefined || expandable) && expanded !== undefined && (
-          <IconButton className="!-ml-2 !mr-4">
+          <IconButton className="!-ml-2 !mr-2">
             {expanded ? <ExpandLess /> : <ExpandMore />}
           </IconButton>
         )}
@@ -144,19 +82,7 @@ export function ChapterHeader({
               {chapter.title}
             </Typography>
             {action}
-            {chapter.description && (
-              <Typography
-                variant="body2"
-                color="text.secondary"
-                onClick={(e) => e.stopPropagation()}
-              >
-                {chapter.description}
-              </Typography>
-            )}
           </div>
-          {chapter.skills.length > 0 && (
-            <div className="flex items-start flex-wrap gap-2">{skillChips}</div>
-          )}
         </div>
       </div>
       {chapter.suggestedEndDate && chapter.suggestedStartDate && (
