+++ components/content-link/ContentLink.tsx
@@ -149,7 +149,7 @@ export function ContentLink({
   const gap = size == "small" ? "gap-2" : "gap-4";
 
   const cursor = !disabled ? "cursor-pointer" : "cursor-default";
-  const frameSize = size == "small" ? "w-10 h-10" : "w-16 h-16";
+  const frameSize = size == "small" ? "w-10 h-10" : "w-12 h-12";
 
   let icon =
     content.__typename === "MediaContent" ? (
@@ -211,7 +211,7 @@ export function ContentLink({
       </div>
       <div className="group-hover:group-enabled:translate-x-0.5">
         <div
-          className={`flex items-center ${
+          className={`flex pb-1 items-center ${
             size == "small" ? "gap-1 -ml-0.5" : "gap-1.5 -ml-1"
           }`}
         >
@@ -221,16 +221,12 @@ export function ContentLink({
               className={"!h-4 px-0 !text-[0.6rem]"}
               label={chip.label}
               sx={{ backgroundColor: chip.color }}
-              classes={{ label: size == "small" ? "!px-2 mt-[0.1rem]" : "" }}
             />
           ))}
         </div>
         <Typography
-          variant="subtitle1"
-          fontSize={size == "small" ? "0.8rem" : "1.25rem"}
-          fontWeight="500"
-          color={disabled ? "text.disabled" : ""}
-          sx={size == "small" ? { lineHeight: 1.5 } : { marginBottom: -0.5 }}
+          variant="body2"
+          color={disabled ? "text.disabled" : "text.secondary"}
         >
           {content.metadata.name}
         </Typography>
